# server/openapi.yaml
openapi: 3.1.0
info:
  title: Trier OS API
  version: 3.5.1
  description: |
    Enterprise Industrial Operating System REST API.
    Single-instance deployment — all plants connect to one corporate server.
    Plant scope is set per-request via the x-plant-id header.
  contact:
    name: Trier OS Engineering
  license:
    name: Proprietary — © 2026 Trier OS. All Rights Reserved.
servers:
  - url: /api
    description: Trier OS corporate server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: query
      name: api_key
  parameters:
    PlantIdHeader:
      name: x-plant-id
      in: header
      required: true
      schema:
        type: string
        pattern: '^[a-zA-Z0-9_-]{1,64}$'
      description: Plant database identifier. Use 'all_sites' for corporate-scoped endpoints.
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /operator-trust/recommendations:
    get:
      summary: "List AI/system recommendations for operators"
      operationId: getoperatorTrustRecommendations
      tags:
        - Operator Trust
      parameters:
        - $ref: '#/components/parameters/PlantIdHeader'
      responses:
        '200':
          description: Successful operation
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '500':
          description: Server error
  /operator-trust/feedback:
    post:
      summary: "Submit operator feedback (accept/reject) on recommendation"
      operationId: postoperatorTrustFeedback
      tags:
        - Operator Trust
      parameters:
        - $ref: '#/components/parameters/PlantIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Successful operation
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '500':
          description: Server error
  /operator-trust/outcomes:
    get:
      summary: "List outcomes of past recommendations"
      operationId: getoperatorTrustOutcomes
      tags:
        - Operator Trust
      parameters:
        - $ref: '#/components/parameters/PlantIdHeader'
      responses:
        '200':
          description: Successful operation
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '500':
          description: Server error
...